+++ crates/web-sys/src/features/gen_Performance.rs
@@ -116,6 +116,7 @@ extern "C" {
     #[doc = ""]
     #[doc = "*This API requires the following crate features to be activated: `Performance`*"]
     pub fn get_entries_by_type(this: &Performance, entry_type: &str) -> ::js_sys::Array;
+    #[cfg(not(web_sys_unstable_apis))]
     # [wasm_bindgen (catch , method , structural , js_class = "Performance" , js_name = mark)]
     #[doc = "The `mark()` method."]
     #[doc = ""]
@@ -123,6 +124,35 @@ extern "C" {
     #[doc = ""]
     #[doc = "*This API requires the following crate features to be activated: `Performance`*"]
     pub fn mark(this: &Performance, mark_name: &str) -> Result<(), JsValue>;
+    #[cfg(web_sys_unstable_apis)]
+    #[cfg(feature = "PerformanceMark")]
+    # [wasm_bindgen (catch , method , structural , js_class = "Performance" , js_name = mark)]
+    #[doc = "The `mark()` method."]
+    #[doc = ""]
+    #[doc = "[MDN Documentation](https://developer.mozilla.org/en-US/docs/Web/API/Performance/mark)"]
+    #[doc = ""]
+    #[doc = "*This API requires the following crate features to be activated: `Performance`, `PerformanceMark`*"]
+    #[doc = ""]
+    #[doc = "*This API is unstable and requires `--cfg=web_sys_unstable_apis` to be activated, as"]
+    #[doc = "[described in the `wasm-bindgen` guide](https://wasm-bindgen.github.io/wasm-bindgen/web-sys/unstable-apis.html)*"]
+    pub fn mark(this: &Performance, mark_name: &str) -> Result<PerformanceMark, JsValue>;
+    #[cfg(web_sys_unstable_apis)]
+    #[cfg(all(feature = "PerformanceMark", feature = "PerformanceMarkOptions",))]
+    # [wasm_bindgen (catch , method , structural , js_class = "Performance" , js_name = mark)]
+    #[doc = "The `mark()` method."]
+    #[doc = ""]
+    #[doc = "[MDN Documentation](https://developer.mozilla.org/en-US/docs/Web/API/Performance/mark)"]
+    #[doc = ""]
+    #[doc = "*This API requires the following crate features to be activated: `Performance`, `PerformanceMark`, `PerformanceMarkOptions`*"]
+    #[doc = ""]
+    #[doc = "*This API is unstable and requires `--cfg=web_sys_unstable_apis` to be activated, as"]
+    #[doc = "[described in the `wasm-bindgen` guide](https://wasm-bindgen.github.io/wasm-bindgen/web-sys/unstable-apis.html)*"]
+    pub fn mark_with_mark_options(
+        this: &Performance,
+        mark_name: &str,
+        mark_options: &PerformanceMarkOptions,
+    ) -> Result<PerformanceMark, JsValue>;
+    #[cfg(not(web_sys_unstable_apis))]
     # [wasm_bindgen (catch , method , structural , js_class = "Performance" , js_name = measure)]
     #[doc = "The `measure()` method."]
     #[doc = ""]
@@ -130,29 +160,92 @@ extern "C" {
     #[doc = ""]
     #[doc = "*This API requires the following crate features to be activated: `Performance`*"]
     pub fn measure(this: &Performance, measure_name: &str) -> Result<(), JsValue>;
+    #[cfg(not(web_sys_unstable_apis))]
     # [wasm_bindgen (catch , method , structural , js_class = "Performance" , js_name = measure)]
     #[doc = "The `measure()` method."]
     #[doc = ""]
     #[doc = "[MDN Documentation](https://developer.mozilla.org/en-US/docs/Web/API/Performance/measure)"]
     #[doc = ""]
     #[doc = "*This API requires the following crate features to be activated: `Performance`*"]
-    pub fn measure_with_start_mark(
+    pub fn measure_with_str(
         this: &Performance,
         measure_name: &str,
         start_mark: &str,
     ) -> Result<(), JsValue>;
+    #[cfg(not(web_sys_unstable_apis))]
     # [wasm_bindgen (catch , method , structural , js_class = "Performance" , js_name = measure)]
     #[doc = "The `measure()` method."]
     #[doc = ""]
     #[doc = "[MDN Documentation](https://developer.mozilla.org/en-US/docs/Web/API/Performance/measure)"]
     #[doc = ""]
     #[doc = "*This API requires the following crate features to be activated: `Performance`*"]
-    pub fn measure_with_start_mark_and_end_mark(
+    pub fn measure_with_str_and_end_mark(
         this: &Performance,
         measure_name: &str,
         start_mark: &str,
         end_mark: &str,
     ) -> Result<(), JsValue>;
+    #[cfg(web_sys_unstable_apis)]
+    #[cfg(feature = "PerformanceMeasure")]
+    # [wasm_bindgen (catch , method , structural , js_class = "Performance" , js_name = measure)]
+    #[doc = "The `measure()` method."]
+    #[doc = ""]
+    #[doc = "[MDN Documentation](https://developer.mozilla.org/en-US/docs/Web/API/Performance/measure)"]
+    #[doc = ""]
+    #[doc = "*This API requires the following crate features to be activated: `Performance`, `PerformanceMeasure`*"]
+    #[doc = ""]
+    #[doc = "*This API is unstable and requires `--cfg=web_sys_unstable_apis` to be activated, as"]
+    #[doc = "[described in the `wasm-bindgen` guide](https://wasm-bindgen.github.io/wasm-bindgen/web-sys/unstable-apis.html)*"]
+    pub fn measure(this: &Performance, measure_name: &str) -> Result<PerformanceMeasure, JsValue>;
+    #[cfg(web_sys_unstable_apis)]
+    #[cfg(feature = "PerformanceMeasure")]
+    # [wasm_bindgen (catch , method , structural , js_class = "Performance" , js_name = measure)]
+    #[doc = "The `measure()` method."]
+    #[doc = ""]
+    #[doc = "[MDN Documentation](https://developer.mozilla.org/en-US/docs/Web/API/Performance/measure)"]
+    #[doc = ""]
+    #[doc = "*This API requires the following crate features to be activated: `Performance`, `PerformanceMeasure`*"]
+    #[doc = ""]
+    #[doc = "*This API is unstable and requires `--cfg=web_sys_unstable_apis` to be activated, as"]
+    #[doc = "[described in the `wasm-bindgen` guide](https://wasm-bindgen.github.io/wasm-bindgen/web-sys/unstable-apis.html)*"]
+    pub fn measure_with_str(
+        this: &Performance,
+        measure_name: &str,
+        start_or_measure_options: &str,
+    ) -> Result<PerformanceMeasure, JsValue>;
+    #[cfg(web_sys_unstable_apis)]
+    #[cfg(feature = "PerformanceMeasure")]
+    # [wasm_bindgen (catch , method , structural , js_class = "Performance" , js_name = measure)]
+    #[doc = "The `measure()` method."]
+    #[doc = ""]
+    #[doc = "[MDN Documentation](https://developer.mozilla.org/en-US/docs/Web/API/Performance/measure)"]
+    #[doc = ""]
+    #[doc = "*This API requires the following crate features to be activated: `Performance`, `PerformanceMeasure`*"]
+    #[doc = ""]
+    #[doc = "*This API is unstable and requires `--cfg=web_sys_unstable_apis` to be activated, as"]
+    #[doc = "[described in the `wasm-bindgen` guide](https://wasm-bindgen.github.io/wasm-bindgen/web-sys/unstable-apis.html)*"]
+    pub fn measure_with_str_and_end_mark(
+        this: &Performance,
+        measure_name: &str,
+        start_or_measure_options: &str,
+        end_mark: &str,
+    ) -> Result<PerformanceMeasure, JsValue>;
+    #[cfg(web_sys_unstable_apis)]
+    #[cfg(all(feature = "PerformanceMeasure", feature = "PerformanceMeasureOptions",))]
+    # [wasm_bindgen (catch , method , structural , js_class = "Performance" , js_name = measure)]
+    #[doc = "The `measure()` method."]
+    #[doc = ""]
+    #[doc = "[MDN Documentation](https://developer.mozilla.org/en-US/docs/Web/API/Performance/measure)"]
+    #[doc = ""]
+    #[doc = "*This API requires the following crate features to be activated: `Performance`, `PerformanceMeasure`, `PerformanceMeasureOptions`*"]
+    #[doc = ""]
+    #[doc = "*This API is unstable and requires `--cfg=web_sys_unstable_apis` to be activated, as"]
+    #[doc = "[described in the `wasm-bindgen` guide](https://wasm-bindgen.github.io/wasm-bindgen/web-sys/unstable-apis.html)*"]
+    pub fn measure_with_performance_measure_options(
+        this: &Performance,
+        measure_name: &str,
+        start_or_measure_options: &PerformanceMeasureOptions,
+    ) -> Result<PerformanceMeasure, JsValue>;
     #[cfg(web_sys_unstable_apis)]
     # [wasm_bindgen (method , structural , js_class = "Performance" , js_name = measureUserAgentSpecificMemory)]
     #[doc = "The `measureUserAgentSpecificMemory()` method."]